hub/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.7
django-cors-headers==3.2.1
django-stubs==1.5.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
googlemaps==4.4.1
idna==2.9
mypy==0.770
mypy-extensions==0.4.3
PyJWT==1.7.1
pytz==2020.1
requests==2.23.0
Shapely==1.7.0
sqlparse==0.3.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9